third_party/WebKit/LayoutTests/external/wpt/css/css-position/position-sticky-top.html

<!DOCTYPE html>
<title>position:sticky elements should respect the top constraint</title>
<link rel="help" href="https://www.w3.org/TR/css-position-3/#sticky-pos" />
<meta name="assert" content="This test checks that position:sticky elements obey their top anchor after scrolling" />

<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>

<style>
.scroller {
  position: relative;
  width: 100px;
  height: 200px;
  overflow: scroll;
}

.contents {
  height: 500px;
}

.prepadding {
  height: 100px;
}

.container {
  height: 200px;
}

.innerpadding {
  height: 50px;
}

.sticky {
  background-color: green;
  position: sticky;
  top: 50px;
}

.box {
  width: 100%;
  height: 100px;
}
</style>

<div id="scroller1" class="scroller">
  <div class="contents">
    <div class="prepadding"></div>
    <div class="container">
      <div class="innerpadding"></div>
      <div class="sticky box"></div>
    </div>
  </div>
</div>

<div id="scroller2" class="scroller">
  <div class="contents">
    <div class="prepadding"></div>
    <div class="container">
      <div class="innerpadding"></div>
      <div class="sticky box"></div>
    </div>
  </div>
</div>

<div id="scroller3" class="scroller">
  <div class="contents">
    <div class="prepadding"></div>
    <div class="container">
      <div class="innerpadding"></div>
      <div class="sticky box"></div>
    </div>
  </div>
</div>

<script>
test(() => {
  const scroller = document.getElementById('scroller1');
  const sticky = scroller.querySelector('.sticky');

  const initialOffset = sticky.offsetTop;
  scroller.scrollTop = 50;
  assert_equals(sticky.offsetTop, initialOffset);
}, 'before reaching the sticking point the sticky box should not be offset');

test(() => {
  const scroller = document.getElementById('scroller2');
  const sticky = scroller.querySelector('.sticky');

  scroller.scrollTop = 125;
  // TODO(smcgruer): I'm not sure this math holds up in all cases; am I assuming
  // something important about the size of the prepadding and innerpadding here?
  //
  // Note that this calculation does already assume that the scrollTop above
  // puts us in the 'stuck' range, which is fine. I am just concerned that it
  // isn't robust for all padding/etc values in the stuck range.
  //
  // It does see to work if I halve or double all the CSS/scroll values, so
  // maybe it's fine and I'm just being paranoid.
  assert_equals(sticky.offsetTop, scroller.scrollTop + 50);
}, 'after reaching the sticking point the sticky box should be offset');

test(() => {
  const scroller = document.getElementById('scroller3');
  const container = scroller.querySelector('.container');
  const sticky = container.querySelector('.sticky');

  scroller.scrollTop = 200;
  const maxOffsetInContainer = container.clientHeight - sticky.clientHeight;
  assert_equals(sticky.offsetTop - container.offsetTop, maxOffsetInContainer);
}, 'the sticky box should not be pushed outside its containing block');
</script>
